requirements.txt
eventlet==0.25.0
python-socketio==4.3.1
Flask==1.1.1
libsass==0.19.2
